add transcoding image optimisation module

+++ derpibooru_dl/parser.py
@@ -2,6 +2,10 @@
 # -*- coding: utf-8 -*-
 
 import urllib.request, json, os, re
+import config
+
+if config.enable_images_optimisations:
+	from . import imgOptimizer
 
 def get_ID_by_URL(URL:str):
 	return URL.split('?')[0].split('/')[-1]
@@ -23,13 +27,45 @@ def download(outdir, data):
 	else:
 		filename=os.path.join(outdir, "{}.{}".format(data["id"],
 			data["original_format"]))
-	if not os.path.isfile(filename):
-		print(filename)
-		print('https:'+os.path.splitext(data['image'])[0]+'.'+data["original_format"])
-		urlstream=urllib.request.urlopen(
-			'https:'+os.path.splitext(data['image'])[0]+'.'+data["original_format"]
+	if config.enable_images_optimisations and \
+		data["original_format"] in set(['png', 'jpg', 'jpeg', 'gif']):
+		if not os.path.isfile(filename) and \
+			not os.path.isfile(os.path.join(outdir, "{} {}.{}".format(
+				data["id"],
+				re.sub('[/\[\]:;|=*".?]', '', os.path.splitext(data["file_name"])[0]),
+				imgOptimizer.getExt[data["original_format"]]
+		))):
+			print(filename)
+			print('https:'+os.path.splitext(data['image'])[0]+'.'+data["original_format"])
+			urlstream=urllib.request.urlopen(
+				'https:'+os.path.splitext(data['image'])[0]+'.'+data["original_format"]
+				)
+			file = open(filename, 'wb')
+			file.write(urlstream.read())
+			urlstream.close()
+			file.close()
+		if not os.path.isfile(os.path.join(outdir, "{} {}.{}".format(
+				data["id"],
+				re.sub('[/\[\]:;|=*".?]', '', os.path.splitext(data["file_name"])[0]),
+				imgOptimizer.getExt[data["original_format"]]
+		))):
+			imgOptimizer.transcode(
+				filename,
+				outdir,
+				"{} {}".format(
+					data["id"],
+					re.sub('[/\[\]:;|=*".?]', '', os.path.splitext(data["file_name"])[0])
+				),
+				data
 			)
-		file = open(filename, 'wb')
-		file.write(urlstream.read())
-		urlstream.close()
-		file.close()
+	else:
+		if not os.path.isfile(filename):
+			print(filename)
+			print('https:'+os.path.splitext(data['image'])[0]+'.'+data["original_format"])
+			urlstream=urllib.request.urlopen(
+				'https:'+os.path.splitext(data['image'])[0]+'.'+data["original_format"]
+				)
+			file = open(filename, 'wb')
+			file.write(urlstream.read())
+			urlstream.close()
+			file.close()
